refactor: AccountTransaction 에서 EventDepositTransaction 분리
