+++ src/main/java/com/github/kellyihyeon/stanceadmin/presentation/accounttransaction/AccountTransactionController.java
@@ -2,7 +2,6 @@ package com.github.kellyihyeon.stanceadmin.presentation.accounttransaction;
 
 import com.github.kellyihyeon.stanceadmin.apis.AccountTransactionApi;
 import com.github.kellyihyeon.stanceadmin.application.accounttransaction.AccountTransactionService;
-import com.github.kellyihyeon.stanceadmin.application.accounttransaction.dto.EventDepositCreation;
 import com.github.kellyihyeon.stanceadmin.application.accounttransaction.dto.MemberShipFeeDepositCreation;
 import com.github.kellyihyeon.stanceadmin.domain.member.MemberType;
 import com.github.kellyihyeon.stanceadmin.models.*;
@@ -49,20 +48,6 @@ public class AccountTransactionController implements AccountTransactionApi {
     @Override
     public ResponseEntity<Void> saveCardPaymentTransaction(CardPayment cardPayment) {
         return null;
-    }
-
-    @Override
-    public ResponseEntity<Void> saveEventDepositTransaction(EventDepositInput input) {
-        EventDepositCreation serviceDto = EventDepositCreation.create(
-                input.getEventId(),
-                input.getDepositorIds(),
-                input.getAmount(),
-                TimeConverter.convertToLocalDate(input.getDepositDate()),
-                input.getDescription()
-        );
-
-        service.saveEventDepositTransaction(serviceDto);
-        return ResponseEntity.created(URI.create("")).build();
     }
 
     @Override
